Pattern added and regex updated
I checked the current regex against the list of VA Code sections from the Title, Chapter, Section Structures csv from the State Decoded. The pattern I added was the only one that the regex did not capture. I updated the regex to capture this last pattern. On to the next state! 