CourtListener links should wrap the citation in quotes in the search page URL
Without quotes, a citation search on "410 U.S. 113" brings back:

    410 U.S. 257, 93 S. Ct. 880, 35 L. Ed. 2d 247, 1973 U.S. LEXIS 113
    (410 U.S. .... LEXIS 113)

and

    507 U.S. 410, 113 S. Ct. 1505, 123 L. Ed. 2d 99, 1993 U.S. LEXIS 2401`
    (U.S. 410, 113)

This would be partially fixed by https://github.com/freelawproject/courtlistener/issues/381,
which would exclude matching if the citation crosses citations. That's
the case in both of these examples. But the quotes also ensure the terms
also occur in the right order, which might be a problem in other examples.